reactor(shorebird_cli): organize platform-specific classes in the platforms dir
