+++ packages/shorebird_cli/lib/src/archive_analysis/archive_analysis.dart
@@ -2,5 +2,4 @@ export 'android_archive_differ.dart';
 export 'apple_archive_differ.dart';
 export 'file_set_diff.dart';
 export 'linux_bundle_differ.dart';
-export 'plist.dart';
 export 'windows_archive_differ.dart';
