+++ packages/shorebird_cli/lib/src/platform/apple/macho.dart
