+++ packages/shorebird_cli/lib/src/platform/apple/plist.dart
