+++ packages/shorebird_cli/lib/src/platform/windows/windows.dart
