+++ packages/shorebird_cli/test/src/platform/android/android_test.dart
@@ -7,7 +7,7 @@ import 'package:shorebird_cli/src/engine_config.dart';
 import 'package:shorebird_cli/src/platform/platform.dart';
 import 'package:test/test.dart';
 
-import '../mocks.dart';
+import '../../mocks.dart';
 
 void main() {
   group('AndroidArch', () {
