+++ packages/shorebird_cli/test/src/platform/apple/apple_test.dart
@@ -13,7 +13,7 @@ import 'package:shorebird_cli/src/shorebird_artifacts.dart';
 import 'package:shorebird_cli/src/shorebird_env.dart';
 import 'package:test/test.dart';
 
-import '../mocks.dart';
+import '../../mocks.dart';
 
 void main() {
   group(InvalidExportOptionsPlistException, () {
