+++ packages/shorebird_cli/test/src/platform/linux/linux_test.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:pub_semver/pub_semver.dart';
-import 'package:shorebird_cli/src/platform/linux.dart';
+import 'package:shorebird_cli/src/platform/linux/linux.dart';
 import 'package:test/test.dart';
 
 void main() {
